+++ packages/react-module-signalr/src/use-signalr-provider.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import { ISignalRProvider, Topic } from '@equinor/fusion-framework-module-signalr';
 
-export const useProviderTopic = <T>(
+export const useSignalRProvider = <T>(
     provider: ISignalRProvider,
     hubId: string,
     topicId: string
