+++ packages/react-module-signalr/src/use-signalr.ts
@@ -1,10 +1,11 @@
 import { useModule } from '@equinor/fusion-framework-react-module';
 import { moduleKey } from '@equinor/fusion-framework-module-signalr';
-import { useProviderTopic } from './use-provider-topic';
+import { useSignalRProvider } from './use-signalr-provider';
 
 export const useTopic = <T>(
     hubId: string,
     topicId: string
-): ReturnType<typeof useProviderTopic<T>> => useProviderTopic(useModule(moduleKey), hubId, topicId);
+): ReturnType<typeof useSignalRProvider<T>> =>
+    useSignalRProvider(useModule(moduleKey), hubId, topicId);
 
 export default useTopic;
